+++ EliteAPI/Events/Engineering.cs
@@ -21,7 +21,7 @@ namespace EliteAPI.Events
         public long Level { get; internal set; }
 
         [JsonProperty("Quality")]
-        public double Quality { get; internal set; }
+        public float Quality { get; internal set; }
 
         [JsonProperty("ExperimentalEffect", NullValueHandling = NullValueHandling.Ignore)]
         public string ExperimentalEffect { get; internal set; }
